- NEW: Implemented unknownString, sepPhased, sepUnphased parameters for allelematrix call. - CHG: Changed a lot of ids to dbIds. - NEW: Added germplasmName to BrapiMarkerProfile. - NEW: Added studies filtering by type. - CHG: Changed datatype of the linkage group max position from int to long.

+++ branches/hackathon2016/src/jhi/brapi/resource/LinkageGroup.java
@@ -7,7 +7,7 @@ public class LinkageGroup
 {
 	private String linkageGroupId;
 	private int numberMarkers;
-	private int maxPosition;
+	private long maxPosition;
 
 	public String getLinkageGroupId()
 	{
@@ -29,12 +29,12 @@ public class LinkageGroup
 		this.numberMarkers = numberMarkers;
 	}
 
-	public int getMaxPosition()
+	public long getMaxPosition()
 	{
 		return maxPosition;
 	}
 
-	public void setMaxPosition(int maxPosition)
+	public void setMaxPosition(long maxPosition)
 	{
 		this.maxPosition = maxPosition;
 	}
